Add constant for subdir permissions within the job queue
When archiving jobs, new directories have to be created, as
jobs are archived in groups of 10000. Add a constant describing
the permissions of these newly created directories.

Note that, due to the type, the constant cannot be part
of src/Ganeti/Constants.hs, as we do not have means to export
it to python. Fortunately, this isn't needed.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>