Add script that checks the copyrights of a dirty local repo
Script to check whether the local dirty commits are changing files
which do not have an updated copyright.

The script will determine your current remote branch and local branch,
from which it will extract the commits to analyze.  Afterwards, for
each commit, it will see which files are being modified and, for each
file, it will check the copyright.

Just call the script before pushing your changes and before
devel/review.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>