Another workaround for GHC linking error "unknown symbol"
This error appears after adding the dependency to lifted-base.
See also #683 and ee7caf27.

I surmise that something like this is happening: Some code, that uses
the library, is only referenced through Template Haskell. A probable bug
in GHC linker fails to notice it and doesn't include the library when
linking, causing the error. Referencing the library explicitly in THH.hs
solves the problem.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>