Consider job-IDs queried for twice only once
As reading jobs from disk is an expensive operation, when querying
for jobs, we optimize by considering which values the job-id is asked
for in the filter. As any reasonable person would not add the same
clause twice in an Or-clause, the implicit assumption was that the
job ids obtained by inspecting the filter would be unique; hence the
information for those jobs could be returned directly.

That all did hold true and everyone was happy till with commit 13ef1fa5
a query for a single job suddenly returned the job twice violating
WaitForJobChange's expectation. Fix this by taking the nub of the
jids asked for in the filter.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>