Fix watcher tampering with instance userdown QA
The watcher was tampering with the instance user down QA because, when
the watcher sees an instance marked an 'USER_down', it will cleanup
the instance, that is it will stop the instance, thus marking it as
'ADMIN_down'.  The change in instance status was causing the instance
user down QA to fail sporadically.  This patch pauses the watcher
during this test.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Reviewed-by: Helga Velroyen <helgav@google.com>