Make gnt-debug delay interruptible
The gnt-debug delay command could be useful as a means of acquiring
locks for testing purposes. In practice, to be useful it should be
interruptible, otherwise we risk race conditions or long delays.

This patch follows the examples of the move-instance command and the
gnt-debug test-jobqueue commands, and introduces a mechanism for
communicating with the delay command, allowing it to be interrupted
early.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>