Refactor move-instance QA test
The move-instance QA test will have to be changed in the following
patches to allow testing opportunistic locking.

This patch retains the same functionality as before, but allows an
iallocator to be used, and splits the move back and forth into two
invocations, the first of which will be wrapped later on.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>