Configure luxid to run with the same user as masterd
Currently masterd and luxid run under separate user accounts, which
makes it impossible for them to share access to the job queue.
This patch adds the easiest fix - to configure luxid to run under the
same user as masterd.

It's open for discussion if we should keep the distinction between luxid
and masterd users/groups at the code level (constants) as now, or to
merge these constants together.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>