Add logging functions lifted into MonadBase
Calling logging functions in a monad transformer stack requires
instances for all the transformers.
This patch adds a module with functions named just as in Ganeti.Logging,
but lifted into MonadBase. This allows using them in any transformer
stack which has MonadBase instance and whose base monad is MonadLog.
This condition is satisfied for all IO-based stacks.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>