A quick fix for broken 'make distcheck'
The offending target 'test/hs/offline-test.sh' requires the dependencies
even when the test isn't actually run. This is because currently our
Makefile doesn't distinguish between prerequisites for building test
scripts and running them.

This is a quick fix that hides the dependencies if the Haskell tests
aren't built, but it doesn't address the core problem. This will be
fixed in a separate patch series.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>