Add reason for job pickup to the trail
Add a new entry in the reason trail when a job is picked up by MasterD from the
hard drive, after LuxiD put it there.

Note that the signature of NameToReasonSrc is changed in an incompatible way,
although it's a public method because in this commit we also change its only
usage in the codebase.

Signed-off-by: Michele Tartara <mtartara@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>