Instance create with OS image
Extend 'LUInstanceCreate' to image the instance's first disk if an OS
image is specified via the OS params.  If the OS image is a file, it
will be copied to the node via SSH.  If ths OS image is a URL, it will
be passed directly to the node, which will then download the file.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>