Introduce class for temporary disk creation
The instance zeroing requires that a virtual appliance capable of
performing it is used, and short of provisioning another VM, the best
Ganeti can do is reuse the virtualization environment of the instance.
To do this, a disk is needed to host the OS performing the zeroing, and
this patch introduces a utility function that temporarily creates one.

This disk is not meant to last, and almost all Ganeti checks will
scream and flail upon seeing it. This is intentional. It should not
survive any more than needed, and the context manager syntax should
enforce its use as such.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>