Rename compression option in gnt-backup export
The gnt-backup export option --compress did not do what most would
expect upon seeing the name: compress the exported image. Instead, it
used compression to try and speed up the transfer of the image,
decompressing it prior to writing the file. To reduce confusion, this
patch renames the option to --transport-compression. The change can
still be done in 2.11 as this option was introduced in this version
only, and it is beneficial as it allows a proper --compress or
--image-compression option to be added at a later time.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Michele Tartara <mtartara@google.com>