Factor out and alter the instance running check
Examining whether an instance is running is useful outside the instance
state check, especially as the check relies on the admin_state recorded
within the configuration, which the code treats as the desired rather
than the actual state. This patch pulls the check out into a separate
function. It also updates the function to account for possible user
shutdowns, if the right option is passed.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>