Add a transitional configuration lock
The lock is used for protecting concurrent operations on the
configuration on the Python side.

The lock must be last wrt to the lock ordering so that it can be always
obtained regardless of other locks held by a client. During operations
that hold this lock, no other locks are ever acquired.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>