BGL is not a set lock
Despite what the name suggests, the Big Ganeti Lock does not act as a
set of all locks. It is just a lock as any other lock from the point
of view of the Ganeti lock hierarchy.

It's special meaning comes from the fact that all jobs, except for
lock-less queries, need to acquire it at least in shared mode. This
requirement, however, is only verified at mcpu. As that property was
never verified outside mcpu in the past, we're not making anything
worse by not verifying it in wconfd.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>