Add a mock for the wconfd module
As our so-called unit tests for the logical units are more
like integration tests, calling the actual LU through an almost
live mcpu, we need a quite involved "mock" for the wconfd module,
called indirectly through mcpu. The difficulties arise from the
fact that both, the mcpu and the LUs verify lock ownership at
various places.

To keep things simple, we only keep track of the mocked state
of affairs where the result is actually checked. With more
consistency checks in the LUs, this module has to be implemented
further.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>